Restructure for test of all transforms
