+++ lib/transformerTemplate.js
