+++ transformers/default.js
@@ -1,2 +1,2 @@
-var template = require('./template');
+var template = require('../lib/transformerTemplate');
 module.exports = template('ADR_LABEL', 'ADR_CITY', 'ADR_STATE', 'ADR_ZIP5')
